Research and testing on animations
Did some research on animations and failed to figure out how to get the current animation name. I have a lead on doing it by entering the stylesheets document and get/ changing the name from there.

I originally thought you could use the element you used the animation on, but that proved to not be the case.