Animation switching by time of day
implemented a function that compares two times. The current time of day and two given times in order to decide which portion of the day it currently is. This function cleans up the if statements in my animationSwitcher function and allows me to automatically switch between animations. I've decided to do this between format, because the exact switching one would only work if it was checking in the back end since a user can make the switcher "miss a beat" when in an alert window and stay on the wrong animation after it should've switched.