put on hiatus
I finished implementing the formulas from the wiki, yet all my results are mostly wrong and at this point I've decided to throw in the towel for this one. I've been crunching on solving how the solar noon formula worked and it seems like I've lost too much time to this. Might revisit this in the future once I've gained some more experience in Js, but for now this is it for this project. See you in the next one!